Move container module out of base package
The base package should have no direct dependencies to feature module to avoid circular dependency issues.
