+++ packages/client/src/container-modules.ts
@@ -30,27 +30,27 @@ import {
     routingModule,
     zorderModule
 } from '~glsp-sprotty';
-import glspBoundsModule from '../features/bounds/di.config';
-import glspCommandPaletteModule from '../features/command-palette/di.config';
-import glspContextMenuModule from '../features/context-menu/di.config';
-import { glspServerCopyPasteModule } from '../features/copy-paste/di.config';
-import glspDecorationModule from '../features/decoration/di.config';
-import glspEditLabelModule from '../features/edit-label/di.config';
-import glspExportModule from '../features/export/di.config';
-import modelHintsModule from '../features/hints/di.config';
-import glspHoverModule from '../features/hover/di.config';
-import layoutModule from '../features/layout/di.config';
-import glspMouseToolModule from '../features/mouse-tool/di.config';
-import navigationModule from '../features/navigation/di.config';
-import glspSelectModule from '../features/select/di.config';
-import sourceModelWatcherModule from '../features/source-model-watcher/di.config';
-import svgMetadataModule from '../features/svg-metadata/di.config';
-import toolFeedbackModule from '../features/tool-feedback/di.config';
-import toolPaletteModule from '../features/tool-palette/di.config';
-import { enableDefaultToolsOnFocusLossModule, toolsModule } from '../features/tools/di.config';
-import { markerNavigatorModule, validationModule } from '../features/validation/di.config';
-import glspViewportModule from '../features/viewport/di.config';
-import defaultGLSPModule from './di.config';
+import defaultGLSPModule from './base/di.config';
+import glspBoundsModule from './features/bounds/di.config';
+import glspCommandPaletteModule from './features/command-palette/di.config';
+import glspContextMenuModule from './features/context-menu/di.config';
+import { glspServerCopyPasteModule } from './features/copy-paste/di.config';
+import glspDecorationModule from './features/decoration/di.config';
+import glspEditLabelModule from './features/edit-label/di.config';
+import glspExportModule from './features/export/di.config';
+import modelHintsModule from './features/hints/di.config';
+import glspHoverModule from './features/hover/di.config';
+import layoutModule from './features/layout/di.config';
+import glspMouseToolModule from './features/mouse-tool/di.config';
+import navigationModule from './features/navigation/di.config';
+import glspSelectModule from './features/select/di.config';
+import sourceModelWatcherModule from './features/source-model-watcher/di.config';
+import svgMetadataModule from './features/svg-metadata/di.config';
+import toolFeedbackModule from './features/tool-feedback/di.config';
+import toolPaletteModule from './features/tool-palette/di.config';
+import { enableDefaultToolsOnFocusLossModule, toolsModule } from './features/tools/di.config';
+import { markerNavigatorModule, validationModule } from './features/validation/di.config';
+import glspViewportModule from './features/viewport/di.config';
 
 export const DEFAULT_MODULES = [
     defaultModule,
